:refactor: recycle: Separate in components header and container of the resultSimulationScree
