+++ app/src/main/java/com/example/ifplanmilk/ui/components/simulation/result/ResultSimulationHeader.kt
@@ -1,0 +1,50 @@
+package com.example.ifplanmilk.ui.components.simulation.result
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.ifplanmilk.ui.theme.IFPlanMilkTheme
+
+@Composable
+fun ResultSimulationHeader(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit = {}
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(bottom = 4.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(text = "Resultado da simulação", modifier.weight(1f))
+        TextButton(onClick = {onClick()}) {
+            Row(
+                horizontalArrangement = Arrangement.spacedBy(4.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(text = "Finalizar")
+                Icon(imageVector = Icons.Filled.Check, contentDescription = null)
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun ResultSimulationHeaderPreview() {
+    IFPlanMilkTheme {
+        ResultSimulationHeader()
+    }
+}
