+++ app/src/main/java/com/example/ifplanmilk/ui/screens/simulation/result/ResultSimulationScreen.kt
@@ -1,6 +1,5 @@
 package com.example.ifplanmilk.ui.screens.simulation.result
 
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -11,16 +10,11 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.sharp.KeyboardArrowUp
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedCard
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -35,9 +29,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.ifplanmilk.data.model.IFPlanSeeItem
 import com.example.ifplanmilk.ui.components.button.IFPlanButton
-import com.example.ifplanmilk.ui.components.form.IFPlanSeeItemResult
+import com.example.ifplanmilk.ui.components.simulation.result.ResultSimulationContainer
+import com.example.ifplanmilk.ui.components.simulation.result.ResultSimulationHeader
 import com.example.ifplanmilk.ui.screens.simulation.animal.AnimalSimulationUiState
 import com.example.ifplanmilk.ui.screens.simulation.area.AreaSimulationUiState
 import com.example.ifplanmilk.ui.screens.simulation.climateSoil.ClimateSoilSimulationUiState
@@ -45,7 +39,6 @@ import com.example.ifplanmilk.ui.screens.simulation.economy.EconomySimulationUiS
 import com.example.ifplanmilk.ui.screens.simulation.sliders.SlidersSimulationScreen
 import com.example.ifplanmilk.ui.screens.simulation.sliders.SlidersSimulationViewModel
 import com.example.ifplanmilk.ui.theme.IFPlanMilkTheme
-import com.example.ifplanmilk.ui.theme.Typography
 
 @Composable
 fun ResultSimulation(
@@ -56,6 +49,7 @@ fun ResultSimulation(
     animalState: AnimalSimulationUiState,
     economyState: EconomySimulationUiState,
     climateSoilState: ClimateSoilSimulationUiState,
+    onNavigateToHome: () -> Unit = {}
 ) {
     var showBottomSheet by remember { mutableStateOf(false) }
 
@@ -63,78 +57,6 @@ fun ResultSimulation(
     val slidersState by slidersViewModel.uiState.collectAsStateWithLifecycle()
 
     val result = uiState.resultSimulation
-    val soilAtAnimalsResult = listOf(
-        IFPlanSeeItem(
-            title = "Tensão da água no solo (bar)",
-            value = result?.tenAguaSolo ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Produção de forragem (kg MV/m2)",
-            value = result?.prodForragem ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Capacidade de suporte (animais)",
-            value = result?.capaSuporte ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Taxa de lotação (vacas/ha)",
-            value = result?.taxaLotacao ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "ITU",
-            value = result?.itu ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "DPL (L/vaca/dia)",
-            value = result?.dpl ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Pegada hídrica (L H2O/L leite)",
-            value = result?.pegadaHidrica ?: 0.0
-        ),
-    )
-    val systemsEconomicResult = listOf(
-        IFPlanSeeItem(
-            title = "Produção diária (L/dia)",
-            value = result?.prodDiaria ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Produção de leite (L/ha/dia)",
-            value = result?.prodLeiteDia ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Produção de leite (L/ha/ano)",
-            value = result?.prodLeiteAno ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Perda receita estresse (R$/ano)",
-            value = result?.perdaReceitaEstresse ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "COE (R$/L)",
-            value = result?.coe ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "COT (R$/L)",
-            value = result?.cot ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "ML (R$/L)",
-            value = result?.mlArea ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Receita por área (R$/ha/ano)",
-            value = result?.receitaTotalAno ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "TRCI (%a.a.)",
-            value = result?.trci ?: 0.0
-        ),
-        IFPlanSeeItem(
-            title = "Payback (anos)",
-            value = result?.payback ?: 0.0
-        ),
-    )
 
     LaunchedEffect(slidersState) {
         onEvent(
@@ -158,51 +80,14 @@ fun ResultSimulation(
             modifier = Modifier.weight(1f),
             verticalArrangement = Arrangement.spacedBy(16.dp),
         ) {
-            Text(text = "Resultado da simulação", modifier.fillMaxWidth())
+            ResultSimulationHeader(
+                onClick = { onNavigateToHome() }
+            )
 
-            Column(
-                modifier = Modifier
-                    .verticalScroll(rememberScrollState())
-                    .padding(bottom = 26.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp),
-            ) {
-
-                OutlinedCard(
-                    modifier = modifier.fillMaxWidth(),
-//            onClick = onClick,
-                    colors = CardDefaults.outlinedCardColors(),
-                    border = BorderStroke(2.dp, MaterialTheme.colorScheme.inverseSurface)
-                ) {
-                    Text(
-                        text = "Solo-Água-Planta-Animal",
-                        style = Typography.headlineSmall,
-                        modifier = Modifier.padding(8.dp)
-                    )
-                    Column(modifier = Modifier.padding(8.dp)) {
-                        soilAtAnimalsResult.forEach { item ->
-                            IFPlanSeeItemResult(item = item)
-                        }
-                    }
-                }
-
-                OutlinedCard(
-                    modifier = modifier.fillMaxWidth(),
-//            onClick = onClick,
-                    colors = CardDefaults.outlinedCardColors(),
-                    border = BorderStroke(2.dp, MaterialTheme.colorScheme.inverseSurface)
-                ) {
-                    Text(
-                        text = "Sistemas-Custos-Resultado-Econômico",
-                        style = Typography.headlineSmall,
-                        modifier = Modifier.padding(8.dp)
-                    )
-                    Column(modifier = Modifier.padding(8.dp)) {
-                        systemsEconomicResult.forEach { item ->
-                            IFPlanSeeItemResult(item = item)
-                        }
-                    }
-                }
-            }
+            ResultSimulationContainer(
+                modifier = modifier.fillMaxWidth(),
+                result = result
+            )
 
         }
 
@@ -241,6 +126,7 @@ fun ResultSimulation(
             }
         }
     }
+
     if (showBottomSheet) {
         SlidersSimulationScreen(
             onDismiss = { showBottomSheet = false },
